+++ controllers/actions.go
@@ -8,18 +8,80 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/klog"
-
 	"context"
-
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
+	"reflect"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/softonic/rate-limit-operator/api/istio_v1alpha3"
 )
 
-func (r *RateLimitReconciler) applyEnvoyFilter(desired istio_v1alpha3.EnvoyFilter, found *istio_v1alpha3.EnvoyFilter, nameEnvoyFilter string, controllerNamespace string) (ctrl.Result, error) {
+func (r *RateLimitReconciler) decomissionk8sObjectResources()  error {
+
+
+	err := r.decomissionEnvoyFilters(r.EnvoyFilters)
+	if err != nil {
+		klog.Infof("Cannot remove EFs %v. Error %v",r.EnvoyFilters , err)
+	}
+
+	err = r.decomissionConfigMapRatelimit(r.configMapRateLimit)
+	if err != nil {
+		klog.Infof("Cannot remove EFs %v. Error %v",r.configMapRateLimit , err)
+	}
+
+
+	return nil
+}
+
+
+func (r *RateLimitReconciler) decomissionEnvoyFilters(EnvoyFilters []*istio_v1alpha3.EnvoyFilter )  error {
+
+
+	for _,envoyfilter := range EnvoyFilters {
+		err := r.deleteEnvoyFilter(*envoyfilter)
+		if err != nil {
+			return err
+			klog.Infof("Cannot remove EF %v. Error %v",envoyfilter , err)
+		}
+	}
+
+	return nil
+
+}
+
+
+
+func (r *RateLimitReconciler) decomissionConfigMapRatelimit(configMapRateLimit v1.ConfigMap )  error {
+
+	err := r.deleteConfigMap(configMapRateLimit)
+	if err != nil {
+		return err
+		klog.Infof("Cannot remove ConfigMap %v. Error %v",configMapRateLimit , err)
+	}
+
+	return nil
+
+}
+
+
+func ( r *RateLimitReconciler) decomissionDeploymentVolumes(sources []v1.VolumeProjection, volumes []v1.Volume) error {
+
+	err := r.removeVolumeFromDeployment(r.DeploymentRL, sources, volumes)
+	if err != nil {
+		klog.Infof("Cannot remove VolumeSource from deploy %v. Error %v", r.DeploymentRL, err)
+		return err
+	}
+
+	err = r.Update(context.TODO(), r.DeploymentRL)
+	if err != nil {
+		klog.Infof("Cannot Update Deployment %s. Error %v", "istio-system-ratelimit", err)
+		return err
+	}
+
+}
+
+func (r *RateLimitReconciler) applyEnvoyFilter(desired istio_v1alpha3.EnvoyFilter, found *istio_v1alpha3.EnvoyFilter, nameEnvoyFilter string, controllerNamespace string) (error) {
 
 	err := r.Get(context.TODO(), types.NamespacedName{
 		Namespace: controllerNamespace,
@@ -30,7 +92,7 @@ func (r *RateLimitReconciler) applyEnvoyFilter(desired istio_v1alpha3.EnvoyFilte
 		err = r.Create(context.TODO(), &desired)
 		if err != nil {
 			klog.Infof("Cannot Create EnvoyFilter %s. Error %v", desired.Name, err)
-			return ctrl.Result{}, err
+			return err
 		}
 	} else {
 
@@ -39,12 +101,12 @@ func (r *RateLimitReconciler) applyEnvoyFilter(desired istio_v1alpha3.EnvoyFilte
 		err = r.Patch(context.TODO(), &desired, client.Apply, applyOpts...)
 		if err != nil {
 			klog.Infof("Cannot Patch EnvoyFilter %s. Error %v", desired.Name, err)
-			return ctrl.Result{}, err
+			return err
 		}
-		return ctrl.Result{}, nil
+		return nil
 	}
 
-	return ctrl.Result{}, nil
+	return nil
 
 }
 
@@ -60,7 +122,38 @@ func (r *RateLimitReconciler) deleteEnvoyFilter(envoyFilter istio_v1alpha3.Envoy
 
 }
 
-func (r *RateLimitReconciler) createDesiredConfigMap(rateLimitInstance *networkingv1alpha1.RateLimit, desiredNamespace string, name string) (v1.ConfigMap, error) {
+func (r *RateLimitReconciler) CreateOrUpdateConfigMap(rateLimitInstance *networkingv1alpha1.RateLimit, controllerNamespace string, baseName string) error {
+
+	configmapDesired, err := r.generateConfigMap(rateLimitInstance, controllerNamespace, baseName)
+	if err != nil {
+		klog.Infof("Cannot create %v, Error: %v", configmapDesired, err)
+		return err
+	}
+
+	found := v1.ConfigMap{}
+
+	r.configMapRateLimit, err = r.getConfigMap(baseName, controllerNamespace)
+	if err != nil {
+		err = r.Create(context.TODO(), &configmapDesired)
+		if err != nil {
+			//return ctrl.Result{}, client.IgnoreNotFound(err)
+			klog.Infof("Cannot create %v, Error: %v", configmapDesired, err)
+		}
+	} else if !reflect.DeepEqual(configmapDesired, found) {
+
+		applyOpts := []client.PatchOption{client.ForceOwnership, client.FieldOwner("rate-limit-controller")}
+
+		err = r.Patch(context.TODO(), &configmapDesired, client.Apply, applyOpts...)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+
+}
+
+func (r *RateLimitReconciler) generateConfigMap(rateLimitInstance *networkingv1alpha1.RateLimit, desiredNamespace string, name string) (v1.ConfigMap, error) {
 
 	configMapData := make(map[string]string)
 
@@ -109,6 +202,24 @@ func (r *RateLimitReconciler) deleteConfigMap(configMapRateLimit v1.ConfigMap) e
 	err := r.Delete(context.TODO(), &configMapRateLimit)
 	if err != nil {
 		klog.Infof("Cannot delete ConfigMap %s. Error %v", configMapRateLimit.Name, err)
+		return err
+	}
+
+	return nil
+
+}
+
+func (r *RateLimitReconciler) UpdateDeployment(sources []v1.VolumeProjection, volumes []v1.Volume) error {
+
+	err := r.addVolumeFromDeployment(r.DeploymentRL, sources, volumes)
+	if err != nil {
+		klog.Infof("Cannot add VolumeSource from deploy %v. Error %v", r.DeploymentRL, err)
+		return err
+	}
+
+	err = r.Update(context.TODO(), r.DeploymentRL)
+	if err != nil {
+		klog.Infof("Cannot Update Deployment %s. Error %v", "istio-system-ratelimit", err)
 		return err
 	}
